src/proxy/requirements.txt
paho-mqtt==1.6.1
influxdb_client==1.30.0
Flask==2.1.3
pymongo==4.2.0
waitress==2.1.2